console/requirements.txt
python-socketio[client]==5.5.2
eventlet==0.33
pyqrcode==1.2.1
